UIIN-955: Init work on performance improvements
This PR includes some initial work related to performance improvements:

- Cleanup unnecessary manifests which caused same requests executing multiple times.
- Refactored routing to a single InstanceRoute and moved filter segment into a regular query param. This allowed for removing some code duplication around items route, holdings route and instances route and it will simplify introducing new routes in the future.
- Added memo to InstancesView to help with unnecessary re-renders.
