--- src/components/HoldingFilters/index.js
@@ -1,1 +1,0 @@
-export { default } from './HoldingFilters';
