+++ src/components/HoldingsRecordFilters/index.js
@@ -1,0 +1,2 @@
+export { default } from './HoldingsRecordFilters';
+export { default as holdingsRecordFilterRenderer } from './holdingsRecordFilterRenderer';
