+++ src/components/InstanceFilters/index.js
@@ -1,1 +1,2 @@
 export { default } from './InstanceFilters';
+export { default as instanceFilterRenderer } from './instanceFilterRenderer';
