+++ src/components/ItemFilters/index.js
@@ -1,1 +1,2 @@
 export { default } from './ItemFilters';
+export { default as itemFilterRenderer } from './itemFilterRenderer';
