+++ src/components/index.js
@@ -1,6 +1,6 @@
 export { default as FilterNavigation } from './FilterNavigation';
 export { default as RepeatableField } from './RepeatableField';
-export { default as InstanceFilters } from './InstanceFilters';
-export { default as HoldingFilters } from './HoldingFilters';
-export { default as ItemFilters } from './ItemFilters';
+export { default as InstanceFilters, instanceFilterRenderer } from './InstanceFilters';
+export { default as HoldingsRecordFilters, holdingsRecordFilterRenderer } from './HoldingsRecordFilters';
+export { default as ItemFilters, itemFilterRenderer } from './ItemFilters';
 export { default as InstancesList } from './InstancesList';
