+++ src/filterConfig.js
@@ -1,0 +1,157 @@
+import {
+  instanceFilterRenderer,
+  holdingsRecordFilterRenderer,
+  itemFilterRenderer,
+} from './components';
+
+export const instanceFilterConfig = [
+  {
+    name: 'effectiveLocation',
+    cql: 'item.effectiveLocationId',
+    values: [],
+  },
+  {
+    name: 'language',
+    cql: 'languages',
+    values: [],
+  },
+  {
+    name: 'resource',
+    cql: 'instanceTypeId',
+    values: [],
+  },
+  {
+    name: 'mode',
+    cql: 'modeOfIssuanceId',
+    values: [],
+  },
+  {
+    name: 'location',
+    cql: 'holdingsRecords.permanentLocationId',
+    values: [],
+  },
+  {
+    name: 'staffSuppress',
+    cql: 'staffSuppress',
+    values: [],
+  },
+  {
+    name: 'discoverySuppress',
+    cql: 'discoverySuppress',
+    values: [],
+  },
+];
+
+export const instanceIndexes = [
+  { label: 'ui-inventory.search.all', value: 'all', queryTemplate: 'keyword all "%{query.query}"' },
+  { label: 'ui-inventory.contributor', value: 'contributor', queryTemplate: 'contributors =/@name "%{query.query}"' },
+  { label: 'ui-inventory.title', value: 'title', queryTemplate: 'title all "%{query.query}"' },
+  { label: 'ui-inventory.identifierAll', value: 'identifier', queryTemplate: 'identifiers =/@value "%{query.query}"' },
+  { label: 'ui-inventory.isbn', prefix: '- ', value: 'isbn', queryTemplate: 'identifiers =/@value/@identifierTypeId="<%= identifierTypeId %>" "%{query.query}"' },
+  { label: 'ui-inventory.issn', prefix: '- ', value: 'issn', queryTemplate: 'identifiers =/@value/@identifierTypeId="<%= identifierTypeId %>" "%{query.query}"' },
+  { label: 'ui-inventory.subject', value: 'subject', queryTemplate: 'subjects="%{query.query}"' },
+  // { label: 'ui-inventory.barcode', value: 'item.barcode', queryTemplate: 'item.barcode=="%{query.query}"' },
+  { label: 'ui-inventory.instanceHrid', value: 'hrid', queryTemplate: 'hrid="%{query.query}"' },
+  { label: 'ui-inventory.instanceId', value: 'id', queryTemplate: 'id="%{query.query}"' },
+  { label: 'ui-inventory.querySearch', value: 'querySearch', queryTemplate: '%{query.query}' },
+];
+
+export const instanceSortMap = {
+  Title: 'title',
+  publishers: 'publication',
+  Contributors: 'contributors',
+};
+
+export const holdingIndexes = [
+  { label: 'ui-inventory.search.all', value: 'all', queryTemplate: 'keyword all "%{query.query}"' },
+  { label: 'ui-inventory.isbn', value: 'isbn', queryTemplate: 'identifiers =/@value/@identifierTypeId="<%= identifierTypeId %>" "%{query.query}"' },
+  { label: 'ui-inventory.issn', value: 'issn', queryTemplate: 'identifiers =/@value/@identifierTypeId="<%= identifierTypeId %>" "%{query.query}"' },
+  { label: 'ui-inventory.holdingsHrid', value: 'hrid', queryTemplate: 'holdingsRecords.hrid="%{query.query}"' },
+  { label: 'ui-inventory.querySearch', value: 'querySearch', queryTemplate: '%{query.query}' },
+];
+
+export const holdingSortMap = {};
+
+export const holdingFilterConfig = [
+  {
+    name: 'effectiveLocation',
+    cql: 'item.effectiveLocationId',
+    values: [],
+  },
+  {
+    name: 'holdingsPermanentLocation',
+    cql: 'holdingsRecords.permanentLocationId',
+    values: [],
+  },
+  {
+    name: 'discoverySuppress',
+    cql: 'holdingsRecords.discoverySuppress',
+    values: [],
+  },
+];
+
+export const itemIndexes = [
+  { label: 'ui-inventory.search.all', value: 'all', queryTemplate: 'keyword all "%{query.query}"' },
+  { label: 'ui-inventory.barcode', value: 'item.barcode', queryTemplate: 'item.barcode=="%{query.query}"' },
+  { label: 'ui-inventory.isbn', value: 'isbn', queryTemplate: 'identifiers =/@value/@identifierTypeId="<%= identifierTypeId %>" "%{query.query}"' },
+  { label: 'ui-inventory.issn', value: 'issn', queryTemplate: 'identifiers =/@value/@identifierTypeId="<%= identifierTypeId %>" "%{query.query}"' },
+  { label: 'ui-inventory.itemHrid', value: 'hrid', queryTemplate: 'item.hrid="%{query.query}"' },
+  { label: 'ui-inventory.querySearch', value: 'querySearch', queryTemplate: '%{query.query}' },
+];
+
+export const itemFilterConfig = [
+  {
+    name: 'materialType',
+    cql: 'item.materialTypeId',
+    values: [],
+  },
+  {
+    name: 'itemStatus',
+    cql: 'item.status.name',
+    values: [],
+  },
+  {
+    name: 'effectiveLocation',
+    cql: 'item.effectiveLocationId',
+    values: [],
+  },
+  {
+    name: 'holdingsPermanentLocation',
+    cql: 'holdingsRecords.permanentLocationId',
+    values: [],
+  },
+  {
+    name: 'discoverySuppress',
+    cql: 'item.discoverySuppress',
+    values: [],
+  }
+];
+
+export const itemSortMap = {
+  Title: 'title',
+  publishers: 'publication',
+  Contributors: 'contributors',
+};
+
+const config = {
+  instances: {
+    filters: instanceFilterConfig,
+    indexes: instanceIndexes,
+    sortMap: instanceSortMap,
+    renderer: instanceFilterRenderer,
+  },
+  holdings: {
+    filters: holdingFilterConfig,
+    indexes: holdingIndexes,
+    sortMap: holdingSortMap,
+    renderer: holdingsRecordFilterRenderer,
+  },
+  items: {
+    filters: itemFilterConfig,
+    indexes: itemIndexes,
+    sortMap: itemSortMap,
+    renderer: itemFilterRenderer,
+  }
+};
+
+export const getFilterConfig = (segment = 'instances') => config[segment];
