+++ src/routes/index.js
@@ -1,3 +1,2 @@
 export { default as InstancesRoute } from './InstancesRoute';
-export { default as HoldingsRoute } from './HoldingsRoute';
-export { default as ItemsRoute } from './ItemsRoute';
+export default {};
