+++ src/views/index.js
@@ -1,3 +1,3 @@
 export { default as InstancesView } from './InstancesView';
-export { default as HoldingsView } from './HoldingsView';
-export { default as ItemsView } from './ItemsView';
+
+export default {};
