+++ test/bigtest/tests/filters/holdings/effective-location-test.js
@@ -14,7 +14,7 @@ describe('Holdings pane effective location filter', () => {
   const holdingsRoute = new HoldingsRouteInteractor();
 
   beforeEach(function () {
-    this.visit('/inventory/holdings');
+    this.visit('/inventory?segment=holdings');
   });
 
   describe('open effective location filter', () => {
