+++ test/bigtest/tests/filters/items/effectivel-location-test.js
@@ -14,7 +14,7 @@ describe('Item pane effective location filter', () => {
   const itemsRoute = new ItemsRouteInteractor();
 
   beforeEach(function () {
-    this.visit('/inventory/items');
+    this.visit('/inventory?segment=items');
   });
 
   describe('open effective location filter', () => {
