+++ test/bigtest/tests/filters/items/material-type-filter-test.js
@@ -14,7 +14,7 @@ describe('ItemMaterialTypeFilter', () => {
   const itemsRoute = new ItemsRouteInteractor();
 
   beforeEach(function () {
-    this.visit('/inventory/items');
+    this.visit('/inventory?segment=items');
   });
 
   describe('open material type filter', () => {
