+++ test/bigtest/tests/filters/items/permanent-location-filter-test.js
@@ -14,7 +14,7 @@ describe('Item pane permanent location filter', () => {
   const itemsRoute = new ItemsRouteInteractor();
 
   beforeEach(function () {
-    this.visit('/inventory/items');
+    this.visit('/inventory?segment=items');
   });
 
   describe('open permanent location filter', () => {
